Added snapshot testing for boleto component
